Refactor BrowseView and CollectionPictureDisplay to one component
