+++ projects/bp-gallery/src/components/views/browse/BrowseView.scss
